Extreme Value Bounds with Double Sampling with Sensitivity

This function yields extreme value bounds under the assumption that the outcomes of 1-delta of the missing second-round units are ignorable, that is, that they are drawn from an unknown distribution with mean and variance equal to the observed second-round groups.

estimator_ds_sens(Y, Z, R1, Attempt, R2, minY, maxY, delta, strata = NULL,
  alpha = 0.05, data)

Arguments

Y
The (unquoted) outcome variable. Must be numeric.

Z
The (unquoted) assignment indicator variable. Must be numeric and take values 0 or 1.

R1
The (unquoted) initial sample respose indicator variable. Must be numeric and take values 0 or 1.

Attempt
The (unquoted) follow-up sample attempt indicator variable. Must be numeric and take values 0 or 1.

R2
The (unquoted) follow-up sample respose indicator variable. Must be numeric and take values 0 or 1.

minY
The minimum possible value of the outcome (Y) variable.

maxY
The maximum possible value of the outcome (Y) variable.

strata
A single (unquoted) variable that indicates which strata units are in.

alpha
The desired significance level. 0.05 by default.

data
A dataframe

sims
Number of points at which to evaluate sensitivity test. Defaults to 100

Value

A list containing a ggplot object, a dataframe of simulated bounds and cis, and a value of pstar, if it exists.